backend/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13" # Vercel currently only supports 3.12
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx[http2]>=0.28.1",
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]


[project.scripts]
app = "index:app"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.6",
    "pre-commit>=4.0.0",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade - automatically detects deprecated Python patterns
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
    "FURB",   # refurb - modernization rules
]

ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # FastAPI Depends() in defaults is the correct pattern
]

[tool.ruff.lint.pyupgrade]
# Keep runtime typing for Python 3.9+ features
keep-runtime-typing = false
